solve sleuth 601-4
This time we will catch Fry by comparing both distance from the crimes and dates of sightings.
If she was within less than 52 pixels of any of the crimes within no more than 1 days of their occurrence then the details
should be pushed to the list of possible matches with the following format.

{ suspect_x: 0, suspect_y: 0 ,crime_x: 0, crime_y: 0, victimName: "John_Doe" }